requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.0
pyspark==3.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
prometheus-client==0.19.0
pyyaml==6.0.1
loguru==0.7.2
aiofiles==23.2.1
aiohttp==3.9.1